+++ Sources/SentrySwift/SwiftDescriptor.swift
@@ -5,10 +5,7 @@ public class SwiftDescriptor: NSObject {
  
     @objc
     public static func getDescription(_ object: AnyObject) -> String {
-        if let objClass = object as? AnyClass {
-            return String(describing: objClass)
-        }
-        return String(describing: object)
+        return String(describing: type(of: object))
     }
     
 }